Folder restructure
Folder restructure to make flask loading more streamlined and easier to deploy.
