+++ static/styles/HeaderStyles.css
@@ -1,0 +1,45 @@
+/* header.css */
+
+:root {
+    --card-bg: #FFFFFF;
+}
+
+.page-header {
+    display: flex;
+    align-items: center;
+    padding: 10px 20px;
+    background: var(--card-bg);
+    border-bottom: 1px solid #CCCCCC;
+}
+
+.logo {
+    width: 100px;
+    margin-right: 20px;
+}
+
+.page-title {
+    font-size: 48px;
+    color: #000;
+}
+
+.nav-bar {
+    display: flex;
+    justify-content: space-around;
+    background-color: var(--primary-green);
+    padding: 12px 0;
+}
+
+.nav-item {
+    color: var(--card-bg);
+    font-size: 20px;
+    font-weight: 500;
+    cursor: pointer;
+    padding: 8px 16px;
+    border-radius: 4px;
+    transition: background-color 0.3s ease;
+}
+
+.nav-item:hover {
+    background-color: rgba(255, 255, 255, 0.1);
+}
+
