--- styles/CartPageStyles.css
@@ -1,105 +1,0 @@
-.cart-table-wrapper {
-  margin: 40px auto;
-  max-width: 900px;
-  border: 4px solid var(--primary-green);
-  background-color: white;
-  border-radius: 4px;
-  overflow-x: auto;
-}
-
-.cart-table {
-  width: 100%;
-  border-collapse: collapse;
-}
-
-.cart-table th,
-.cart-table td {
-  border: 1px solid #444;
-  padding: 12px;
-  text-align: center;
-  font-size: 14px;
-}
-
-.product-icon {
-  width: 24px;
-  height: 24px;
-  vertical-align: middle;
-  margin-right: 8px;
-}
-
-.remove-btn {
-  background: none;
-  border: none;
-  font-size: 18px;
-  cursor: pointer;
-}
-
-.shipping-selectors {
-  display: flex;
-  justify-content: center;
-  gap: 10px;
-  margin-top: 20px;
-  flex-wrap: wrap;
-}
-
-.shipping-selectors select {
-  padding: 8px;
-  border-radius: 4px;
-  border: 1px solid var(--border-gray);
-  font-size: 14px;
-}
-
-.order-summary {
-  margin: 30px auto;
-  padding: 20px;
-  max-width: 500px;
-  background-color: #ccc;
-  border: 1px solid #666;
-  border-radius: 6px;
-}
-
-.order-summary p {
-  margin: 8px 0;
-  font-size: 16px;
-}
-
-.order-total {
-  float: right;
-  font-weight: bold;
-}
-
-.cart-actions {
-  display: flex;
-  justify-content: center;
-  gap: 40px;
-  margin: 30px 0 60px;
-}
-
-.shop-more-btn,
-.submit-order-btn {
-  padding: 12px 24px;
-  font-size: 18px;
-  font-weight: 600;
-  border-radius: 6px;
-  border: none;
-  cursor: pointer;
-  transition: background-color 0.3s ease;
-}
-
-.shop-more-btn {
-  background-color: #666;
-  color: white;
-}
-
-.submit-order-btn {
-  background-color: var(--primary-green);
-  color: white;
-}
-
-.submit-order-btn:hover {
-  background-color: #1a4534;
-}
-
-.shop-more-btn:hover {
-  background-color: #4d4d4d;
-}
