--- styles/GlobalStyles.css
@@ -1,28 +1,0 @@
-/* global.css */
-
-:root {
-    --primary-green: #21543F;
-    --background: #E3E3E3;
-    --text-color: #000000;
-    --card-bg: #FFFFFF;
-    --accent: #21543F;
-    --border-gray: #CCCCCC;
-  }
-  
-  body {
-    margin: 0;
-    font-family: 'Inter', sans-serif;
-    background-color: var(--background);
-    color: var(--text-color);
-    user-select: none;
-  }
-  
-  input, textarea, select {
-    user-select: text;
-  }
-
-  a {
-    text-decoration: none;
-  }
-  
-
