--- styles/ProductPageStyles.css
@@ -1,192 +1,0 @@
-.logo {
-  width: 100px;
-  margin-right: 20px;
-}
-
-.page-title {
-  font-size: 48px;
-  color: var(--text-color);
-}
-
-.nav-bar {
-  display: flex;
-  justify-content: space-around;
-  background-color: var(--primary-green);
-  padding: 12px 0;
-}
-
-.nav-item {
-  font-weight: 700;         /* Bold */
-  font-size: 20px;
-  color: var(--card-bg);
-  padding: 8px 16px;
-  border-radius: 4px;
-  cursor: pointer;
-  transition: background-color 0.3s ease;
-}
-
-
-.nav-item:hover {
-  background-color: rgba(255, 255, 255, 0.1);
-}
-
-.main-content {
-  padding: 20px 40px;
-}
-
-.customization-info {
-  background-color: #f2f9f6;
-  border-left: 6px solid #21543F;
-  padding: 16px 24px;
-  margin-bottom: 24px;
-  border-radius: 6px;
-  box-shadow: 0 2px 6px rgba(0, 0, 0, 0.05);
-}
-
-.customization-info h2 {
-  font-size: 24px;
-  color: #21543F;
-  margin: 0 0 8px 0;
-}
-
-.customization-info p {
-  font-size: 14px;
-  color: #333;
-  line-height: 1.5;
-  margin: 0;
-}
-
-
-.military-icons {
-  display: flex;
-  justify-content: center;
-  gap: 15px;
-  margin: 20px 0;
-  flex-wrap: wrap;
-}
-
-.military-icons img {
-  width: 60px;
-  height: 60px;
-  object-fit: contain;
-  background-color: var(--card-bg);
-  border-radius: 6px;
-  padding: 5px;
-  box-shadow: 0 0 4px rgba(0, 0, 0, 0.05);
-}
-
-.product-grid {
-  display: flex;
-  flex-wrap: wrap;
-  gap: 20px;
-  justify-content: center;
-}
-
-.product-card {
-  width: 350px;               
-  height: 120px;            
-  overflow: hidden;        
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  gap: 12px;
-  padding: 12px;
-  background: var(--card-bg);
-  border-radius: 8px;
-  box-shadow: 0 2px 6px rgba(0, 0, 0, 0.05);
-  box-sizing: border-box;
-}
-
-
-.product-image {
-  width: 80px;
-  height: 80px;
-  object-fit: contain;
-  border-radius: 4px;
-  flex-shrink: 0;
-}
-
-.product-details {
-  flex-grow: 1;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  gap: 6px;
-}
-
-.product-title-price {
-  display: flex;
-  align-items: center;
-  gap: 8px;
-  font-size: 14px;
-  font-weight: 600;
-}
-
-.separator {
-  color: #888;
-}
-
-.customization-row {
-  display: flex;
-  align-items: center;
-  gap: 6px;
-}
-
-.customization-row select {
-  font-size: 13px;
-  padding: 4px;
-  height: 32px;
-  border-radius: 4px;
-  border: 1px solid var(--border-gray);
-  box-sizing: border-box;
-}
-
-
-.qty-field {
-  font-size: 13px;
-  width: 32px;
-  height: 32px;
-  padding: 0;
-  text-align: center;
-  border-radius: 4px;
-  border: 1px solid var(--border-gray);
-  box-sizing: border-box;
-}
-
-.add-to-cart {
-  display: flex;
-  justify-content: flex-end;
-  margin-top: 30px;
-}
-
-.add-to-cart button {
-  background-color: var(--accent);
-  color: var(--card-bg);
-  font-size: 18px;
-  font-weight: 600;
-  padding: 12px 24px;
-  border: none;
-  border-radius: 6px;
-  cursor: pointer;
-  transition: background-color 0.3s ease;
-}
-
-.add-to-cart-btn {
-  background-color: #21543F;      /* Matches navbar */
-  color: white;
-  border: none;
-  padding: 6px 12px;
-  border-radius: 6px;
-  font-weight: 600;
-  font-size: 14px;
-  cursor: pointer;
-  transition: background-color 0.3s ease;
-  width: 100%;                    /* Optional: full-width button */
-  margin-top: 6px;
-}
-
-.add-to-cart-btn:hover {
-  background-color: #1a4534;      /* Slightly darker on hover */
-}
-
-
