+++ templates/CartPage.html
@@ -1,0 +1,77 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+  <meta charset="UTF-8" />
+  <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+  <title>My Cart</title>
+  <link rel="stylesheet" href="styles/GlobalStyles.css" />
+  <link rel="stylesheet" href="styles/HeaderStyles.css" />
+  <link rel="stylesheet" href="styles/CartPageStyles.css" />
+  
+</head>
+<body>
+
+  <!-- Header -->
+  <div class="page-header">
+    <img src="https://placehold.co/100x100?text=Logo" alt="Logo" class="logo" />
+    <h1 class="page-title">My Cart</h1>
+  </div>
+
+  <!-- Navigation Bar -->
+  <div class="nav-bar">
+    <a href="ProductPage.html" class="nav-item">Products Page</a>
+    <a href="CartPage.html" class="nav-item">Cart</a>
+    <div class="nav-item">Order Status</div>
+    <div class="nav-item">My Account</div>
+    <div class="nav-item">Log Out</div>
+  </div>
+
+  <!-- Main Content -->
+  <div class="main-content">
+
+    <!-- Cart Table -->
+    <div class="cart-table-wrapper">
+      <table class="cart-table">
+        <thead>
+          <tr>
+            <th>Product</th>
+            <th>Color</th>
+            <th>Customization</th>
+            <th>Quantity</th>
+            <th>Price</th>
+            <th>Remove</th>
+          </tr>
+        </thead>
+        <tbody id="cart-items"></tbody>
+
+
+        
+
+      </table>
+      
+    </div>
+
+    <!-- Shipping Selectors -->
+    <div class="shipping-selectors">
+      <select><option>Select Shipping State</option></select>
+      <select><option>Select Shipping City</option></select>
+      <select><option>Select Shipping Address</option></select>
+    </div>
+
+    <!-- Order Summary -->
+    <div class="order-summary">
+      <p><strong>Order Total:</strong>  <span id="order-total">$0.00</span></p>
+      <p><strong>Selected Delivery Address:</strong></p>
+      <p>7 FRANKFORD AVE BLDG 219 ANNISTON AL 36201-0000</p>
+    </div>
+
+    <!-- Buttons -->
+    <div class="cart-actions">
+      <button class="shop-more-btn">Shop More</button>
+      <button class="submit-order-btn">Submit Order</button>
+    </div>
+
+  </div>
+  <script src="js/cart.js"></script>
+</body>
+</html>
